+++ CHANGELOG.md
@@ -20,6 +20,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Delete and Update commands return response based on UoW response
 * Extra Ef Tools package for SqlServer
 
+## [0.16.6] - 09/05/2022
+
+### Fixed
+
+* Bad Test Fixture with SqlServer (fixes #95)
+
 ## [0.16.5] - 08/18/2022
 
 ### Fixed